Use babel-loose to avoid unperformed for of
For of arrays transpiled with babel have a slowdown of 6x (https://kpdecker.github.io/six-speed/).
This might be a dramatic slow down for a performance measurement (in the case of knight tour increases a single run from 36min to 216min!).
Therefore use babel loose that has a better performance but does not cover all es6 semantics (no issue for this project)